requirements.txt
alembic==0.9.6
aniso8601==1.3.0
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.4
chardet==3.0.4
click==6.7
contextlib2==0.5.5
coverage==4.5.1
docopt==0.6.2
Flask==0.12.2
Flask-Cors==3.0.3
Flask-JWT==0.3.2
Flask-JWT-Extended==3.3.4
Flask-Login==0.4.0
Flask-Migrate==2.1.1
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
future==0.16.0
google-api-core==1.3.0
google-auth==1.5.0
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
more-itertools==4.3.0
mysql-connector==2.1.4
paho-mqtt==1.3.1
pathlib2==2.3.2
pbr==4.2.0
pluggy==0.7.1
protobuf==3.6.0
py==1.5.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyJWT==1.4.2
pykwalify==1.6.1
pytest==3.7.1
python-box==3.2.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
PyYAML==3.13
requests==2.19.1
rsa==3.4.2
send-email==201212
simplejson==3.16.0
six==1.11.0
SQLAlchemy==1.1.14
stevedore==1.29.0
tavern==0.17.2
ts==0.4.3
urllib3==1.23
Werkzeug==0.12.2
WTForms==2.1
